requirements.txt
python-box==7.0.1
reusables==0.9.6
unstructured==0.10.0
ctransformers==0.2.21 
pypdf==3.15.1